initial commit of scripts for generating files for submission to GCB
supports Tier 1 output, except for ancillary output

Remapped output includes fractional cells at land/sea interface. It
is unresolved if GCB is okay with this, or if they prefer cells that
are either all ocean or all land. I'm holding off on generating grid
information on the remapped grid until this is resolved.